refactor(migration): move last of upload_hdc.js to migration.js
